Create youtube project.py
YouTube-Data-Harvesting-and-Warehousing
This is my intermediate level Python Project to harvest YouTube data using YouTube Data API and store the data in a MongoDB database as a data lake.After that the data is migrated from the data lake to a SQL database as tables and are displayed in the streamlit application.

##Project##

Introduction:

YouTube is the world's most popular video-sharing platform, with over 2 billion active users. It is a valuable source of data for businesses, researchers, and individuals. This project will demonstrate how to harvest and warehouse YouTube data using SQL, MongoDB, and Streamlit.

Data Collection:

The first step is to collect data from YouTube. This can be done using the YouTube Data API. The API provides access to a wide range of data, including channel information, video statistics, and viewer engagement metrics.

Data Storage:

The collected data can be stored in a variety of ways. In this project, we will use MongoDB and SQL. MongoDB is a NoSQL database that is well-suited for storing large amounts of unstructured data. SQL is a relational database that is well-suited for querying and analyzing structured data.

Data Warehousing:

The data can be stored in a data warehouse, which is a centralized repository for data. Data warehouses are used to store large amounts of data from a variety of sources. They can be used to analyze data, identify trends, and make predictions.

Data Analysis:

The data can be analyzed using a variety of tools. In this project, we will use Streamlit. Streamlit is a Python library that can be used to create interactive web applications. We will use Streamlit to create a dashboard that allows users to visualize and analyze the data.

Conclusion:

This project has demonstrated how to harvest and warehouse YouTube data using SQL, MongoDB, and Streamlit. This approach can be used to collect, store, and analyze data from a variety of sources.


Thank You